clean up idm check
